refactor massively removing tiledb, dask, napari
