+++ requirements.txt
@@ -4,13 +4,10 @@ openslide-python==1.1.2
 pyecvl==1.1.0
 pyeddl==1.2.0
 clize==4.2.1
-dask[complete]==2.20.0
 zarr==2.8.3
 dataclasses==0.6
-tiledb==0.7.0
 progress==1.6
 stringcase==1.2.0
-napari-lazy-openslide==0.2.0
 scikit-image==0.17.2
 fsspec>=0.3.3
 onnx==1.11.0
