+++ slaid/commons/base.py
@@ -2,23 +2,13 @@ import abc
 import inspect
 import logging
 import math
-import os
-import re
 import sys
 from dataclasses import dataclass
-from datetime import datetime as dt
 from enum import Enum
-from pathlib import Path
 from typing import Dict, List, Tuple
 
 import numpy as np
 import PIL
-import tiledb
-import zarr
-from napari_lazy_openslide import OpenSlideStore
-from napari_lazy_openslide.store import ArgumentError, init_attrs
-from skimage.util import view_as_blocks
-from zarr.storage import init_array, init_group
 
 DEFAULT_TILESIZE = 1024
 logger = logging.getLogger()
@@ -407,204 +397,6 @@ class BasicSlideArray(SlideArray):
 
     def _is_channel_first(self):
         return self.image_info.channel == ImageInfo.Channel.FIRST
-
-
-class NapariSlide(BasicSlide):
-
-    def __init__(self, store: "SlideStore", image_info: ImageInfo = None):
-        self._store = store
-        self._slide = store.slide
-        self.image_info = image_info if image_info else self.IMAGE_INFO
-        grp = zarr.open(store, mode="r")
-        multiscales = grp.attrs["multiscales"][0]
-        self._pyramid = [
-            self._create_slide(d) for d in multiscales["datasets"]
-        ]
-
-    @property
-    def masks(self):
-        return self._slide.masks
-
-    def _create_slide(self, dataset):
-        return NapariSlideArray(self._read_from_store(dataset),
-                                self._slide.IMAGE_INFO)
-
-    def _read_from_store(self, dataset):
-        return zarr.open(store=self._store, path=dataset["path"], mode='r')
-        # FIXME
-        #  import dask.array as da
-        #  logger.debug('dataset %s', dataset)
-        #  return da.from_zarr(self._store, component=dataset["path"])
-
-    def __getitem__(self, key) -> "SlideArray":
-        return self._pyramid[key]
-
-    @property
-    def dimensions(self) -> Tuple[int, int]:
-        return self._slide.dimensions
-
-    @property
-    def filename(self):
-        return self._slide.filename
-
-    def read_region(self, location: Tuple[int, int], level: int,
-                    size: Tuple[int, int]) -> Image:
-        return self._slide.read_region(location, level, size)
-
-    def get_best_level_for_downsample(self, downsample: int):
-        return self._slide.get_best_level_for_downsample(downsample)
-
-    @property
-    def level_dimensions(self):
-        return self._slide.level_dimensions
-
-    @property
-    def level_downsamples(self):
-        return self._slide.level_downsamples
-
-
-class NapariSlideArray:
-
-    def __init__(self, array: np.ndarray, image_info: ImageInfo = None):
-        self._array = array
-        self._image_info = image_info
-
-    def __getitem__(self, key) -> "SlideArray":
-        if self._is_channel_first():
-            array = self._array[:, key[0], key[1]] if isinstance(
-                key, tuple) else self._array[:, key]
-        else:
-            array = self._array[key[0], key[1], :] if isinstance(
-                key, tuple) else self._array[key, :]
-        return self.__class__(array, self._image_info)
-
-    @property
-    def array(self):
-        return np.array(self._array)
-
-    def _is_channel_first(self):
-        return self._image_info.channel == ImageInfo.Channel.FIRST
-
-    @property
-    def size(self) -> Tuple[int, int]:
-        return self._array.shape[1:] if self._is_channel_first(
-        ) else self._array.shape[:2]
-
-    def reshape(self, shape) -> "SlideArray":
-
-        array = self._array.reshape((3, ) + shape) if self._is_channel_first(
-        ) else self._array.reshape(shape + (3, ))
-        return self.__class__(array, self._image_info)
-
-    def convert(self, image_info: ImageInfo) -> "SlideArray":
-        array = self._image_info.convert(self.array, image_info)
-        return self.__class__(array, image_info)
-
-    def apply_filter(self, filter_: Filter) -> "SlideArray":
-        if self._is_channel_first():
-            array = self.array[:, filter_.array]
-        else:
-            array = self.array[filter_.array, :]
-
-        return self.__class__(array, self._image_info)
-
-    def get_blocks(self, block_shape: Tuple[int,
-                                            int]) -> Tuple[np.ndarray, bool]:
-        if self._is_channel_first():
-            block_shape = (3, ) + block_shape
-            channel_first = True
-        else:
-
-            block_shape = block_shape + (3, )
-            channel_first = False
-        array = view_as_blocks(self.array, block_shape)
-        return array, channel_first
-
-
-def _create_metastore(slide: BasicSlide, tilesize: int) -> Dict[str, bytes]:
-    """Creates a dict containing the zarr metadata
-    for the multiscale openslide image."""
-    store = dict()
-    root_attrs = {
-        "multiscales": [{
-            "name":
-            Path(slide.filename).name,
-            "datasets": [{
-                "path": str(i)
-            } for i in range(slide.level_count)],
-            "version":
-            "0.1",
-        }]
-    }
-    init_group(store)
-    init_attrs(store, root_attrs)
-    for i, (x, y) in enumerate(slide.level_dimensions):
-        init_array(
-            store,
-            path=str(i),
-            shape=(3, y, x) if slide.IMAGE_INFO.channel
-            == ImageInfo.Channel.FIRST else (y, x, 3),
-            chunks=(3, tilesize, tilesize) if slide.IMAGE_INFO.channel
-            == ImageInfo.Channel.FIRST else (tilesize, tilesize, 3),
-            dtype="|u1",
-            compressor=None,
-        )
-    return store
-
-
-class SlideStore(OpenSlideStore):
-
-    def __init__(self, slide: "Slide", tilesize: int = None):
-        self._path = slide.filename
-        self._slide = slide
-        self._tilesize = tilesize or DEFAULT_TILESIZE
-        self._store = _create_metastore(self._slide, self._tilesize)
-        #  self._image_info = image_info if image_info else ImageInfo()
-
-    @property
-    def tilesize(self):
-        return self._tilesize
-
-    @tilesize.setter
-    def tilesize(self, value):
-        self._tilesize = value
-        self._store = _create_metastore(self._slide, value)
-
-    @property
-    def slide(self):
-        return self._slide
-
-    def __getitem__(self, key: str):
-        if key in self._store:
-            # key is for metadata
-            return self._store[key]
-
-        # key should now be a path to an array chunk
-        # e.g '3/4.5.0' -> '<level>/<chunk_key>'
-        try:
-            x, y, level = self._parse_chunk_path(key)
-            location = self._ref_pos(x, y, level)
-            size = (self._tilesize, self._tilesize)
-            logger.debug(
-                'key %s, x %s, y %s, location %s, size %s, dimension %s', key,
-                x, y, location, size, self._slide.level_dimensions[level])
-            tile = self._slide.read_region(location, level, size)
-        except ArgumentError as err:
-            # Can occur if trying to read a closed slide
-            raise err
-        except Exception as ex:
-            logger.error('ex %s', ex)
-            raise KeyError(key) from ex
-
-        return tile.to_array().tobytes()
-
-    def _parse_chunk_path(self, path: str):
-        """Returns x,y chunk coords and pyramid level from string key"""
-        level, ckey = path.split("/")
-        _, y, x = map(int, ckey.split("."))
-        if self._slide.IMAGE_INFO.channel == ImageInfo.Channel.LAST:
-            y, x, _ = _, y, x
-        return x, y, int(level)
 
 
 class ArrayFactory(abc.ABC):
