+++ slaid/commons/openslide.py
@@ -1,37 +1,6 @@
-# NAPARI LAZY OPENSLIDE
-#  Copyright (c) 2020, Trevor Manz
-#  All rights reserved.
-#
-#  Redistribution and use in source and binary forms, with or without
-#  modification, are permitted provided that the following conditions are met:
-#
-#  * Redistributions of source code must retain the above copyright notice, this
-#    list of conditions and the following disclaimer.
-#
-#  * Redistributions in binary form must reproduce the above copyright notice,
-#    this list of conditions and the following disclaimer in the documentation
-#    and/or other materials provided with the distribution.
-#
-#  * Neither the name of napari-lazy-openslide nor the names of its
-#    contributors may be used to endorse or promote products derived from
-#    this software without specific prior written permission.
-#
-#  THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
-#  AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
-#  IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
-#  DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
-#  FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
-#  DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
-#  SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
-#  CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
-#  OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
-#  OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-
 from typing import Tuple
 
 import numpy as np
-import zarr
-from napari_lazy_openslide import OpenSlideStore
 from openslide import open_slide
 from PIL import Image as PIL_Image
 
@@ -92,16 +61,3 @@ class BasicSlide(BaseSlide):
     @property
     def level_downsamples(self):
         return self._level_downsamples
-
-    def to_array(self, level):
-        store = OpenSlideStore(self.filename)
-        grp = zarr.open(store, mode="r")
-        datasets = grp.attrs["multiscales"][0]["datasets"]
-
-        pyramid = [grp.get(d["path"]) for d in datasets]
-        return pyramid[level]
-
-
-def load(filename: str):
-    slide = Slide(filename)
-    return slide
