+++ tests/test_classifiers.py
@@ -7,12 +7,10 @@ from slaid.classifiers.fixed_batch import (FilteredPatchClassifier,
                                            FilteredPixelClassifier,
                                            PixelClassifier)
 from slaid.commons import Mask
-from slaid.commons.base import Filter, Slide, ImageInfo
-from slaid.commons.dask import init_client
+from slaid.commons.base import Filter, Slide
 from slaid.commons.ecvl import BasicSlide as EcvlSlide
 from slaid.commons.openslide import BasicSlide as OpenSlide
-from slaid.models.eddl import TissueModel, TumorModel
-from slaid.writers.zarr import ZarrStorage
+from slaid.writers.zarr_io import ZarrStorage
 from tests.commons import EddlGreenPatchModel, GreenModel
